fixed warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/module-gnu-java/trunk@4745 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ src/gnu-java-module.cpp
@@ -358,7 +358,7 @@ void QoreJavaClassMap::doConstructors(QoreClass &qc, java::lang::Class *jc, Exce
       // get constructor methods
       JArray<java::lang::reflect::Constructor *> *methods = jc->getDeclaredConstructors(false);
 
-      for (int i = 0; i < methods->length; ++i) {
+      for (size_t i = 0; i < (size_t)methods->length; ++i) {
 	 java::lang::reflect::Constructor *m = elements(methods)[i];
 
 #ifdef DEBUG
@@ -403,7 +403,7 @@ void QoreJavaClassMap::doMethods(QoreClass &qc, java::lang::Class *jc, Exception
 
    try {
       JArray<java::lang::reflect::Method *> *methods = jc->getDeclaredMethods();
-      for (int i = 0; i < methods->length; ++i) {
+      for (size_t i = 0; i < (size_t)methods->length; ++i) {
 	 java::lang::reflect::Method *m = elements(methods)[i];
 
 	 QoreString mname;
